dashboard/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
watchdog==5.0.0
pyyaml==6.0.2
python-frontmatter==1.1.0
